requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-MySQL==1.4.0
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
protobuf==3.11.1
PyMySQL==0.9.3
six==1.13.0
Werkzeug==0.16.0
WTForms==2.2.1
